refactor(providers): fold process and superset into providers
`@sidecar/process` was 106 lines with two consumers and `@sidecar/superset`
was a provider adapter in a package of its own that already depended on
`@sidecar/providers` and reached six of its fs/sqlite helpers through the
public barrel. Both become directories of `providers`, and Superset reaches
those helpers by relative path rather than re-entering its own barrel.
`sign-in-stage` keeps a door of its own, because preload and renderer code
reads it and the barrel reaches `node:child_process` and `node:sqlite`.

The invocation helpers stay off the barrel: both callers now live inside the
package, so exporting them would only advertise a process spawner to callers
who have no business holding one.

`INVOCATION_FAILURE.TIMED_OUT` and `OUTPUT_LIMIT` go. No caller ever read
either, and both say the same thing the remaining `FAILED` says — the
invocation produced no bounded answer.
