--- packages/process/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@sidecar/process",
-  "version": "0.5.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js"
-  },
-  "types": "./src/index.ts",
-  "scripts": {
-    "test": "tsx --test 'src/**/*.test.ts'",
-    "typecheck": "tsc --project tsconfig.json"
-  },
-  "devDependencies": {
-    "@types/node": "26.2.0",
-    "tsx": "4.23.12",
-    "typescript": "7.0.2"
-  }
-}
