+++ packages/providers/package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "exports": {
     ".": "./src/index.js",
+    "./superset/sign-in-stage": "./src/superset/sign-in-stage.js",
     "./testing": "./src/testing/index.js"
   },
   "types": "./src/index.ts",
@@ -15,7 +16,6 @@
   "dependencies": {
     "@sidecar/acts": "workspace:*",
     "@sidecar/credentials": "workspace:*",
-    "@sidecar/process": "workspace:*",
     "@sidecar/runtime-contracts": "workspace:*",
     "@sidecar/session": "workspace:*",
     "@sidecar/wire": "workspace:*"
