+++ packages/providers/src/index.ts
@@ -45,3 +45,25 @@ export {
   type WorkspaceHostRegistrationOptions,
   workspaceHostRegistrations,
 } from "./shared/workspace-hosts.js";
+export type { SupersetCommandRunner, SupersetQueryRunner } from "./superset/cli.js";
+export { SupersetCli, type SupersetCliOptions } from "./superset/cli.js";
+export {
+  type SupersetPlugin,
+  type SupersetPluginOptions,
+  supersetPlugin,
+} from "./superset/plugin.js";
+export { supersetHostState } from "./superset/reader.js";
+export {
+  SupersetSignIn,
+  type SupersetSignInOptions,
+  validSupersetSignInCode,
+} from "./superset/sign-in.js";
+export {
+  SUPERSET_SIGN_IN_STAGE,
+  type SupersetOrganizationChoice,
+  type SupersetSignInSnapshot,
+  type SupersetSignInStage,
+} from "./superset/sign-in-stage.js";
+export { type SupersetSnapshot, supersetSnapshot } from "./superset/snapshot.js";
+export { isSupersetControlId, SUPERSET_CONTROL_ID } from "./superset/vocabulary.js";
+export { type SupersetSessionContext, supersetPressedLink } from "./superset/wire.js";
