--- packages/superset/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@sidecar/superset",
-  "version": "0.5.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.js",
-    "./sign-in-stage": "./src/sign-in-stage.js"
-  },
-  "types": "./src/index.ts",
-  "scripts": {
-    "test": "tsx --test 'src/**/*.test.ts'",
-    "typecheck": "tsc --project tsconfig.json"
-  },
-  "devDependencies": {
-    "@types/node": "26.2.0",
-    "tsx": "4.23.12",
-    "typescript": "7.0.2"
-  },
-  "dependencies": {
-    "@sidecar/process": "workspace:*",
-    "@sidecar/providers": "workspace:*",
-    "@sidecar/session": "workspace:*",
-    "@sidecar/wire": "workspace:*"
-  }
-}
